+++ packages/scripts/src/monitor-oa/MonitorAssertions.ts
@@ -1,9 +1,5 @@
-import { Logger } from "@uma/financial-templates-lib";
-import { OptimisticAsserterEthers } from "@uma/contracts-node";
+import { getContractInstanceWithProvider, Logger, MonitoringParams, OptimisticAsserterEthers } from "./common";
 import { logAssertion } from "./MonitorLogger";
-import { getContractInstanceWithProvider } from "../utils/contracts";
-
-import type { MonitoringParams } from "./common";
 
 export async function monitorAssertions(logger: typeof Logger, params: MonitoringParams): Promise<void> {
   const oa = await getContractInstanceWithProvider<OptimisticAsserterEthers>("OptimisticAsserter", params.provider);
