+++ packages/scripts/src/monitor-oa/common.ts
@@ -6,6 +6,10 @@ import { getContractInstanceWithProvider } from "../utils/contracts";
 
 import type { Provider } from "@ethersproject/abstract-provider";
 
+export { OptimisticAsserterEthers } from "@uma/contracts-node";
+export { Logger } from "@uma/financial-templates-lib";
+export { getContractInstanceWithProvider } from "../utils/contracts";
+
 export interface BotModes {
   assertionsEnabled: boolean;
   disputesEnabled: boolean;
